Shapefile

Documentation for Shapefile.

Shapefile.Handle

Shapefile.Interval

Shapefile.MultiPatch

Shapefile.MultiPoint

Shapefile.MultiPointM

Shapefile.MultiPointZ

Shapefile.Point

Shapefile.PointM

Shapefile.PointZ

Shapefile.Polygon

Shapefile.PolygonM

Shapefile.PolygonZ

Shapefile.Polyline

Shapefile.PolylineM

Shapefile.PolylineZ

Shapefile.Rect

Shapefile.Row

Shapefile.Table

Base.iterate

Shapefile.shape

Shapefile.shapes

Shapefile.Handle — Type

Handle

Handle(path::AbstractString, [indexpath::AbstractString])

Load a shapefile into GeoInterface compatible objects. This can be plotted with Plots.jl plot.

The Vector of shape object can be accessed with shapes(handle).

Handle may have a known bounding box, which can be retrieved with GeoInterface.bbox.

Shapefile.Interval — Type

Interval

Represents the range of measures or Z dimension, in a shape file.

Shapefile.MultiPatch — Type

MultiPatch <: GeoInterface.AbstractGeometry

Fields

points: a Vector of Point represents a one or multiple spatial objects.

parts: a Vector of Int32 indicating the object each point belongs to.

parttypes: a Vector of Int32 indicating the type of object each point belongs to.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

Shapefile.MultiPoint — Type

MultiPoint <: GeoInterface.AbstractMultiPoint

Collection of points, from a shape file.

Fields

points: a Vector of Point.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

Shapefile.MultiPointM — Type

MultiPointM <: GeoInterface.AbstractMultiPoint

Collection of points, from a shape file.

Includes a measures field, holding values from each point.

May have a known bounding box, which can be retrieved with GeoInterface.bbox.

Fields

points: a Vector of Point.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

measures: holds values from each point.

Shapefile.MultiPointZ — Type

MultiPointZ <: GeoInterface.AbstractMultiPoint

Fields

points: a Vector of Point.

zvalues: a Vector of Float64 representing the z dimension values.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

measures: holds values from each point.

Shapefile.Point — Type

Point <: GeoInterface.AbstractPoint

Point from a shape file.

Fields x, y hold the spatial location.

Shapefile.PointM — Type

PointM <: GeoInterface.AbstractPoint

Point from a shape file.

Fields x, y hold the spatial location.

Includes a measure field m, holding a value for the point.

Shapefile.PointZ — Type

PointZ <: GeoInterface.AbstractPoint

Three dimensional point, from a shape file.

Fields x, y, z hold the spatial location.

Includes a measure field m, holding a value for the point.

Shapefile.Polygon — Type

Polygon <: GeoInterface.AbstractMultiPolygon

Represents a polygon from a shape file.

Fields

points: a Vector of Point represents a one or multiple closed areas.

parts: a Vector of Int32 indicating the polygon each point belongs to.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

Shapefile.PolygonM — Type

PolygonM <: GeoInterface.AbstractMultiPolygon

Represents a polygon from a shape file

Fields

points: a Vector of Point represents a one or multiple closed areas.

parts: a Vector of Int32 indicating the polygon each point belongs to.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

measures: holds values from each point.

Shapefile.PolygonZ — Type

PolygonZ <: GeoInterface.AbstractMultiPolygon

A three dimensional polygon from a shape file.

Fields

points: a Vector of Point represents a one or multiple closed areas.

parts: a Vector of Int32 indicating the polygon each point belongs to.

zvalues: a Vector of Float64 representing the z dimension values.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

measures: holds values from each point.

Shapefile.Polyline — Type

Polyline <: GeoInterface.AbstractMultiLineString

Represents a single or multiple polylines from a shape file.

Fields

points: a Vector of Point represents a one or multiple lines.

parts: a Vector of Int32 indicating the line each point belongs to.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

Shapefile.PolylineM — Type

PolylineM <: GeoInterface.AbstractMultiLineString

Polyline from a shape file, with measures.

Fields

points: a Vector of Point represents a one or multiple lines.

parts: a Vector of Int32 indicating the line each point belongs to.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

measures: holds values from each point.

Shapefile.PolylineZ — Type

PolylineZ <: GeoInterface.AbstractMultiLineString

Three dimensional polyline of from a shape file.

Fields

points: a Vector of Point represents a one or multiple lines.

parts: a Vector of Int32 indicating the line each point belongs to.

zvalues: a Vector of Float64 representing the z dimension values.

MBR: nothing or the known bounding box. Can be retrieved with GeoInterface.bbox.

measures: holds values from each point.

Shapefile.Rect — Type

Rect

A rectangle object to represent the bounding box for other shape file shapes.

Shapefile.Row — Type

Row

Row(geometry, record::DBFTables.Row)

A struct representing a single record in a shapefile.

Property names accessable by row.x are geometry for the geometry object, and the names of the columns in record.

Shapefile.Table — Type

Table 

Table(path::AbstractString)
Table(shp::Handle{T}, dbf::DBFTables.Table)

Represents both geometries and their associated fields as a Tables.jl compatible table, where each row is is a an object such as a (Point](@ref), Polyline or Polygon.

The Shapefile.Table constructor can read either a String path to a shapefile, or a Shapefile.Handle and a DBFTables.Table.

Property names accessable by table.x are geometry for the geometry object, and the names of the columns in dbf.

Base.iterate — Function

Base.iterate(t::Table)

Iterate over the rows of a Shapefile.Table, yielding a Shapefile.Row for each row.

Shapefile.shape — Method

shapes(row::Row)

Get the geometry associated with a Row from a shapefile Table.

Shapefile.shapes — Method

shapes(t::Table)

Get a vector of the geometries in a shapefile Table, without any metadata.